Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-vision-imageanalysis==1.0.0b3
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.34.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
msrest==0.7.1
numpy==2.3.0
oauthlib==3.2.2
openai==1.98.0
orjson==3.11.1
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
postgrest==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.6.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
selenium==4.30.0
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.3.0
uv==0.8.4
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
SQLAlchemy
psycopg[binary]
alembic
Authlib
docx2txt
pypdf
python-jose[cryptography]